update to be pod -> 1vm 2 pods and refactor util.go
